fix!: change the way cursor is moved
BREAKING: Before this change plugin just returned a string
"{number}{j/k}" and it was executed as a normal command by the remap
itself. This meant that it's not the plugin that moves the cursor, but
the plugin only told nvim the count. I figured that it is not how it
should be and changed it. Because of this change the previous
configuration is now broken.