fix: fixed behaviour with noexpandtab
For files with tabs used as indentation the behaviour was very weird as
the tabs were counted as a single character. This commit changes this
and counts leading tabs as `shiftwidth` characters in order to make the
behaviour more similar to that when using expandtab.

The plugin still does not support wide chars and tabs used in the middle
of a line.